// apps/frontend/src/lib/api/endpoints.schemas.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * astro-react-tsoa-orval-boilerplate
 * OpenAPI spec version: 1.0.0
 */
export interface LoginDto {
  email: string;
  password: string;
}

export interface RegisterDto {
  email: string;
  password: string;
}

export type Me200 = {
  updatedAt: string;
  createdAt: string;
  isDeleted: boolean;
  email: string;
  id: string;
};

/**
 */
export type Login200UserRole =
  (typeof Login200UserRole)[keyof typeof Login200UserRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Login200UserRole = {
  USER: "USER",
} as const;

export type Login200User = {
  updatedAt: string;
  createdAt: string;
  /** */
  role: Login200UserRole;
  isDeleted: boolean;
  password: string;
  email: string;
  id: string;
};

export type Login200 = {
  user: Login200User;
  token: unknown;
};

/**
 */
export type Register200UserRole =
  (typeof Register200UserRole)[keyof typeof Register200UserRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Register200UserRole = {
  USER: "USER",
} as const;

export type Register200User = {
  updatedAt: string;
  createdAt: string;
  /** */
  role: Register200UserRole;
  isDeleted: boolean;
  password: string;
  email: string;
  id: string;
};

export type Register200 = {
  user: Register200User;
  token: unknown;
};